fix: split CLI-orchestrator tests back out of test_publish_hf_ops.py
Rebase side-effect: after #31 renamed test_publish_cli.py → test_publish_hf_ops.py,
git's rename detection merged #30's CLI-orchestrator additions into the HF-ops file.
Restore the intended separation — TestPublishIntegration lives in test_publish_cli.py,
TestCheckRegression and TestPromote stay in test_publish_hf_ops.py.
